+++ legends/__init__.py
@@ -5,6 +5,7 @@
 from legends import utils
 from legends import ui
 from legends.constants import *
+from legends.functions import *
 from legends.gameobjects import *
 from legends.roster import *
 from legends.saveslot import *